requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
Pillow==9.2.0
psycopg2==2.9.3
sqlparse==0.4.3
django-crum==0.7.9
